// src/modules/user/user.model.d.ts
import mongoose, { Document } from "mongoose";
export interface IUser extends Document {
    username: string;
    password: string;
    email: string;
    role: "admin" | "editor" | "manager";
    createdAt: Date;
    updatedAt: Date;
}
export declare const User: mongoose.Model<IUser, {}, {}, {}, mongoose.Document<unknown, {}, IUser, {}, mongoose.DefaultSchemaOptions> & IUser & Required<{
    _id: mongoose.Types.ObjectId;
}> & {
    __v: number;
} & {
    id: string;
}, any, IUser>;
//# sourceMappingURL=user.model.d.ts.map